backend/requirements.txt
fastapi
pydantic
SQLAlchemy
sqlalchemy_utils
sqlmodel
uvicorn
python-dotenv
psycopg2
passlib
pyjwt
datetime
bcrypt
python-multipart
openai
paypalrestsdk
fastapi-mail